pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "oneiro"
version = "0.1.0"
description = "Discord bot for image generation with Huggingface Diffusers"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "jkoelker" }
]
keywords = ["discord", "bot", "diffusers", "image-generation", "stable-diffusion"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "py-cord>=2.4.0",
    "python-dotenv>=1.0.0",
    "torch>=2.0.0",
    "diffusers @ git+https://github.com/huggingface/diffusers.git",
    "transformers>=4.36.0",
    "accelerate>=0.25.0",
    "safetensors>=0.4.0",
    "Pillow>=10.0.0",
    "watchfiles>=0.21.0",
    "bitsandbytes>=0.42.0",
    "peft>=0.7.0",
    "gguf>=0.6.0",
    "kernels>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[project.scripts]
oneiro = "oneiro.main:main"

[project.urls]
Repository = "https://github.com/jkoelker/oneiro"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["oneiro"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
